requirements.txt
APScheduler==3.6.3
attrs==22.1.0
bidict==0.22.0
cachetools==4.2.2
certifi==2022.6.15
click==8.1.3
Flask==2.1.3
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-decouple==3.6
python-engineio==4.3.3
python-socketio==5.7.1
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.39
tomli==2.0.1
tornado==6.2
tzdata==2022.1
tzlocal==4.2
Werkzeug==2.2.1
XlsxWriter==3.0.3
zipp==3.8.1
